v6.6: 25 July Production Deployment

+++ app/mutations/tool_slots/create.rb
@@ -5,9 +5,9 @@ module ToolSlots
     required do
       model   :device, class: Device
       string  :name,   default: "Untitled Slot"
-      integer :x
-      integer :y
-      integer :z
+      float :x
+      float :y
+      float :z
     end
 
     optional do
